pyproject.toml
[project]
name = "quanttradeai"
version = "0.1.0"
description = "A machine learning framework for quantitative trading strategies"
authors = [
    {name = "Akshat Joshi",email = "joshiakshat0511@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "numpy (>=2.3.1,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "yfinance (>=0.2.64,<0.3.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "xgboost (>=3.0.2,<4.0.0)",
    "pandas-ta (>=0.3.14b0,<0.4.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "optuna (>=4.4.0,<5.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "alpha-vantage (>=3.0.0,<4.0.0)",
    "litellm (>=1.75.0,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "pybreaker (>=1.4.0,<2.0.0)",
    "asyncio-throttle (>=1.0.2,<2.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)"
]

[tool.poetry]
packages = [
    { include = "quanttradeai" }
]

[project.scripts]
quanttradeai = "quanttradeai.main:main"
quanttradeai-pipeline = "quanttradeai.main:run_pipeline"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
flake8 = "^7.3.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
